Add files via upload
There are still some bugs that I'm working out. I'm uploading to let you guys see the approach I used. Let me know if you have better ideas or fixes.

Some of it is hard coded for the example code because doing otherwise would require a lot of additional work. If needed we can add it in later once I get it working.